Align transfer operation parameters
